backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.9
pypdf==4.3.1
faiss-cpu==1.13.2
numpy==2.0.1
openai==1.40.6
httpx==0.27.2
sqlalchemy==2.0.32
pytest==8.3.2
